Refactor: 패키지 구조 MVC 외 comm으로 통합
