+++ src/comm/outputMessage.js
